+++ CHANGES.md
@@ -1,3 +1,7 @@
+# Changes - 2019-06-12
+- Added ALUI
+- Modified newgent to support string class types.
+
 # Changes - 2019-06-06
 - Renamed newentity of the game interface to newgameentity
 - Renamed deleteentity of the game interface to deletegameentity